DigitalVolumePhaseOnlyCorrelation

Documentation for DigitalVolumePhaseOnlyCorrelation.

DigitalVolumePhaseOnlyCorrelation.BartlettApodization

DigitalVolumePhaseOnlyCorrelation.BartlettHannApodization

DigitalVolumePhaseOnlyCorrelation.BlackmanApodization

DigitalVolumePhaseOnlyCorrelation.CosineApodization

DigitalVolumePhaseOnlyCorrelation.GaussianApodization

DigitalVolumePhaseOnlyCorrelation.HammingApodization

DigitalVolumePhaseOnlyCorrelation.HanningApodization

DigitalVolumePhaseOnlyCorrelation.KaiserApodization

DigitalVolumePhaseOnlyCorrelation.LanczosApodization

DigitalVolumePhaseOnlyCorrelation.RectApodization

DigitalVolumePhaseOnlyCorrelation.TriangApodization

DigitalVolumePhaseOnlyCorrelation.TukeyApodization

DigitalVolumePhaseOnlyCorrelation.poc

DigitalVolumePhaseOnlyCorrelation.BartlettApodization — Type

DigitalVolumePhaseOnlyCorrelation.BartlettApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.BartlettApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see bartlett function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.BartlettApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.BartlettHannApodization — Type

DigitalVolumePhaseOnlyCorrelation.BartlettHannApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.BartlettHannApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see bartlett_hann function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.BartlettHannApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.BlackmanApodization — Type

DigitalVolumePhaseOnlyCorrelation.BlackmanApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.BlackmanApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see blackman function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.BlackmanApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.CosineApodization — Type

DigitalVolumePhaseOnlyCorrelation.CosineApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.CosineApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see cosine function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.CosineApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.GaussianApodization — Type

DigitalVolumePhaseOnlyCorrelation.GaussianApodization(sz, par; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.GaussianApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see gaussian function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.GaussianApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.HammingApodization — Type

DigitalVolumePhaseOnlyCorrelation.HammingApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.HammingApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see hamming function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.HammingApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.HanningApodization — Type

DigitalVolumePhaseOnlyCorrelation.HanningApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.HanningApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see hanning function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.HanningApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.KaiserApodization — Type

DigitalVolumePhaseOnlyCorrelation.KaiserApodization(sz, par; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.KaiserApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see kaiser function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.KaiserApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.LanczosApodization — Type

DigitalVolumePhaseOnlyCorrelation.LanczosApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.LanczosApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see lanczos function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.LanczosApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.RectApodization — Type

DigitalVolumePhaseOnlyCorrelation.RectApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.RectApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see rect function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.RectApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.TriangApodization — Type

DigitalVolumePhaseOnlyCorrelation.TriangApodization(sz; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.TriangApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see triang function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.TriangApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.TukeyApodization — Type

DigitalVolumePhaseOnlyCorrelation.TukeyApodization(sz, par; kwargs...)

Create an apodization function DigitalVolumePhaseOnlyCorrelation.TukeyApodization up to 3 dimensions, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see tukey function in package DSP.jl.

Example

julia> apod = DigitalVolumePhaseOnlyCorrelation.TukeyApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

DigitalVolumePhaseOnlyCorrelation.poc — Method

poc(sig1, sig2)